refactor: merge CoreCursor into Cursor
