+++ src/cursor/aggregation_cursor.ts
@@ -1,6 +1,5 @@
 import { MongoError } from '../error';
-import { Cursor, CursorOptions } from './cursor';
-import { CursorState } from './core_cursor';
+import { Cursor, CursorOptions, CursorState } from './cursor';
 import type { AggregateOperation, AggregateOptions } from '../operations/aggregate';
 import type { Document } from '../bson';
 import type { Sort } from '../operations/find';
