+++ src/cursor/index.ts
@@ -1,4 +1,3 @@
-export { CoreCursor, CursorState } from './core_cursor';
-export { Cursor } from './cursor';
+export { Cursor, CursorState, CursorStream } from './cursor';
 export { CommandCursor } from './command_cursor';
 export { AggregationCursor } from './aggregation_cursor';
