+++ src/sessions.ts
@@ -18,7 +18,7 @@ import {
 import type { Topology } from './sdam/topology';
 import type { MongoClientOptions } from './mongo_client';
 import type { Cursor } from './cursor/cursor';
-import type { CoreCursor } from './cursor/core_cursor';
+// import type { CoreCursor } from './cursor/core_cursor';
 import type { WriteCommandOptions } from './cmap/wire_protocol/write_command';
 import { executeOperation } from './operations/execute_operation';
 import { RunAdminCommandOperation } from './operations/run_command';
@@ -71,7 +71,7 @@ class ClientSession extends EventEmitter {
   clusterTime?: ClusterTime;
   operationTime?: Timestamp;
   explicit: boolean;
-  owner: symbol | CoreCursor;
+  owner: symbol | Cursor; // TODO - change to AbstractCursor
   defaultTransactionOptions: TransactionOptions;
   transaction: Transaction;
   [kServerSession]?: ServerSession;
